+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Master
-
+* allow to pass minimum component size in mm3 to `--remove_small_blobs` argument (before was hard coded to 200mm3)
 
 
 ## Release 2.13.0 (17.03.2026)